refact(139779): separando testes serializer create
